refactor: Change pin type from size_t to uint8_t in PowerSensingInterface

+++ lib/interfaces/include/PowerSensingInterface.h
@@ -5,7 +5,7 @@
 #include "SharedFirmwareTypes.h"
 #include "etl/singleton.h"
 
-using pin = size_t;
+using pin = uint8_t;
 
 namespace pin_power_sensing_params
 {
